+++ dwave/cloud/auth/flows.py
@@ -33,7 +33,8 @@ from dwave.cloud.auth.creds import Credentials
 from dwave.cloud.auth.server import SingleRequestAppServer, RequestCaptureAndRedirectApp
 from dwave.cloud.config.models import ClientConfig
 from dwave.cloud.regions import resolve_endpoints
-from dwave.cloud.utils import pretty_argvalues, default_user_agent
+from dwave.cloud.utils.http import default_user_agent
+from dwave.cloud.utils import pretty_argvalues
 
 __all__ = ['AuthFlow', 'LeapAuthFlow', 'OAuthError']
